+++ js/api.js
@@ -1,6 +1,7 @@
 import { createOfferMarkers } from './map.js';
 import { activateFiltersForm } from './user-form.js';
 import { showAlert } from './util.js';
+import { isEscapeKey, setDefault } from './util.js';
 
 const getOffers = () => {
   fetch('https://26.javascript.pages.academy/keksobooking/data')
@@ -23,12 +24,53 @@ const getOffers = () => {
 
 
 const sendOffer = (formData) => {
-  fetch('https://26.javascript.pages.academy/keksobooking/',
+  fetch('https://26.javascript.pages.academy/keksobookin/',
     {
       method: 'POST',
       body: formData,
     },
-  );
+  )
+    .then ((response) => {
+      if (response.ok) {
+        setDefault();
+        const successMessage = document.querySelector('#success')
+          .content
+          .querySelector('.success')
+          .cloneNode(true);
+        document.body.append(successMessage);
+        document.addEventListener ('click', () => {
+          successMessage.remove();
+        });
+        document.addEventListener ('keydown', (evt) => {
+          if (isEscapeKey(evt)) {
+            evt.preventDefault();
+            successMessage.remove();
+          }
+        });
+      } else {
+        throw new Error;
+      }
+    })
+    .catch (() => {
+      const errorMessage = document.querySelector('#error')
+        .content
+        .querySelector('.error')
+        .cloneNode(true);
+      document.body.append(errorMessage);
+      document.addEventListener ('click', () => {
+        errorMessage.remove();
+      });
+      document.addEventListener ('keydown', (evt) => {
+        if (isEscapeKey(evt)) {
+          evt.preventDefault();
+          errorMessage.remove();
+        }
+      });
+      const errorButton = errorMessage.querySelector('button');
+      errorButton.addEventListener ('click', () => {
+        errorMessage.remove();
+      });
+    });
 };
 
 export {getOffers,sendOffer};
